+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - Fix the release number.
 - Bug: PowerPlants power generation is no longer rounded.
 - Bug: Repair systems power cost is now based on the system size, not on the ship size.
+- Bug: ordnance launchers data was incorrect in the json data file.
+- Bug: launchers now appear in the possible systems that can use fire control.
+- Bug: sensor and fire control installed systems were missign the edit button.
+- Bug: computer cores could be set to invalid sizes.
+- Bug: now ships can have more than 1 computer core.
 
 ## [0.2.1] - 2026-02-04